+++ plugins/cli/plugin.go
@@ -21,7 +21,7 @@ import (
 
 var (
 	// AppVersion version number
-	AppVersion          = "0.5.0"
+	AppVersion          = "0.5.1"
 	LatestGithubVersion = AppVersion
 
 	// AppName app code name