+++ plugins/dashboard/frontend/src/app/misc/Utils.ts
@@ -1,5 +1,0 @@
-
-// replaces all characters with their uppercase representation during input
-export const toInputUppercase = e => {
-    e.target.value = ("" + e.target.value).toUpperCase();
-};