+++ plugins/webapi/transactions.go
@@ -73,71 +73,156 @@ func findTransactions(i interface{}, c *gin.Context, _ <-chan struct{}) {
 		maxResults = query.MaxResults
 	}
 
-	if (len(query.Bundles) + len(query.Addresses) + len(query.Approvees) + len(query.Tags)) > maxResults {
-		e.Error = "Too many bundle, address, approvee or tag hashes. Max. allowed: " + strconv.Itoa(maxResults)
-		c.JSON(http.StatusBadRequest, e)
-		return
-	}
-
-	txHashes := []string{}
-
+	// should return an error in a sane API but unfortunately we need to keep backwards compatibility
 	if len(query.Bundles) == 0 && len(query.Addresses) == 0 && len(query.Approvees) == 0 && len(query.Tags) == 0 {
 		c.JSON(http.StatusOK, FindTransactionsReturn{Hashes: []string{}})
 		return
 	}
 
-	// Searching for transactions that contains the given bundle hash
-	for _, bdl := range query.Bundles {
-		if err := trinary.ValidTrytes(bdl); err != nil {
-			e.Error = fmt.Sprintf("Bundle hash invalid: %s", bdl)
-			c.JSON(http.StatusBadRequest, e)
-			return
-		}
-
-		txHashes = append(txHashes, tangle.GetBundleTransactionHashes(hornet.HashFromHashTrytes(bdl), true, maxResults-len(txHashes)).Trytes()...)
+	if (len(query.Bundles) + len(query.Addresses) + len(query.Approvees) + len(query.Tags)) > maxResults {
+		e.Error = "too many bundle, address, approvee or tag hashes. max. allowed: " + strconv.Itoa(maxResults)
+		c.JSON(http.StatusBadRequest, e)
+		return
 	}
 
-	// Searching for transactions that contains the given address
-	for _, addr := range query.Addresses {
-		if err := address.ValidAddress(addr); err != nil {
-			e.Error = fmt.Sprintf("address hash invalid: %s", addr)
+	queryBundleHashes := make(map[string]struct{})
+	queryApproveeHashes := make(map[string]struct{})
+	queryAddressHashes := make(map[string]struct{})
+	queryTagHashes := make(map[string]struct{})
+
+	// check all queries first
+	for _, bundleTrytes := range query.Bundles {
+		if err := trinary.ValidTrytes(bundleTrytes); err != nil {
+			e.Error = fmt.Sprintf("bundle hash invalid: %s", bundleTrytes)
 			c.JSON(http.StatusBadRequest, e)
 			return
 		}
-
-		if len(addr) == 90 {
-			addr = addr[:81]
-		}
-
-		txHashes = append(txHashes, tangle.GetTransactionHashesForAddress(hornet.HashFromAddressTrytes(addr), query.ValueOnly, true, maxResults-len(txHashes)).Trytes()...)
+		queryBundleHashes[string(hornet.HashFromHashTrytes(bundleTrytes))] = struct{}{}
 	}
 
-	// Searching for all approvers of the given transactions
-	for _, approveeHash := range query.Approvees {
-		if !guards.IsTransactionHash(approveeHash) {
-			e.Error = fmt.Sprintf("Aprovee hash invalid: %s", approveeHash)
+	for _, approveeTrytes := range query.Approvees {
+		if !guards.IsTransactionHash(approveeTrytes) {
+			e.Error = fmt.Sprintf("aprovee hash invalid: %s", approveeTrytes)
 			c.JSON(http.StatusBadRequest, e)
 			return
 		}
-
-		txHashes = append(txHashes, tangle.GetApproverHashes(hornet.HashFromHashTrytes(approveeHash), maxResults-len(txHashes)).Trytes()...)
+		queryApproveeHashes[string(hornet.HashFromHashTrytes(approveeTrytes))] = struct{}{}
 	}
 
-	// Searching for transactions that contain the given tag
-	for _, tag := range query.Tags {
-		if err := trinary.ValidTrytes(tag); err != nil {
-			e.Error = fmt.Sprintf("Tag invalid: %s", tag)
+	for _, addressTrytes := range query.Addresses {
+		if err := address.ValidAddress(addressTrytes); err != nil {
+			e.Error = fmt.Sprintf("address hash invalid: %s", addressTrytes)
 			c.JSON(http.StatusBadRequest, e)
 			return
 		}
+		if len(addressTrytes) == 90 {
+			addressTrytes = addressTrytes[:81]
+		}
+		queryAddressHashes[string(hornet.HashFromHashTrytes(addressTrytes))] = struct{}{}
+	}
 
-		if len(tag) != 27 {
-			e.Error = fmt.Sprintf("Tag invalid length: %s", tag)
+	for _, tagTrytes := range query.Tags {
+		if err := trinary.ValidTrytes(tagTrytes); err != nil {
+			e.Error = fmt.Sprintf("tag invalid: %s", tagTrytes)
 			c.JSON(http.StatusBadRequest, e)
 			return
 		}
+		if len(tagTrytes) != 27 {
+			e.Error = fmt.Sprintf("tag invalid length: %s", tagTrytes)
+			c.JSON(http.StatusBadRequest, e)
+			return
+		}
+		queryTagHashes[string(hornet.HashFromTagTrytes(tagTrytes))] = struct{}{}
+	}
 
-		txHashes = append(txHashes, tangle.GetTagHashes(hornet.HashFromTagTrytes(tag), true, maxResults-len(txHashes)).Trytes()...)
+	results := make(map[string]struct{})
+	searchedBefore := false
+
+	// search txs by bundle hash
+	for bundleHash := range queryBundleHashes {
+		for _, r := range tangle.GetBundleTransactionHashes(hornet.Hash(bundleHash), true, maxResults-len(results)) {
+			results[string(r)] = struct{}{}
+		}
+		searchedBefore = true
+	}
+
+	if !searchedBefore {
+		// search txs by approvees
+		for approveeHash := range queryApproveeHashes {
+			for _, r := range tangle.GetApproverHashes(hornet.Hash(approveeHash), maxResults-len(results)) {
+				results[string(r)] = struct{}{}
+			}
+		}
+		searchedBefore = true
+	} else {
+		// check if results match at least one of the approvee search criteria
+		for txHash := range results {
+			contains := false
+			for approveeHash := range queryApproveeHashes {
+				if tangle.ContainsApprover(hornet.Hash(approveeHash), hornet.Hash(txHash)) {
+					contains = true
+					break
+				}
+			}
+			if !contains {
+				delete(results, txHash)
+			}
+		}
+	}
+
+	if !searchedBefore {
+		// search txs by address
+		for addressHash := range queryAddressHashes {
+			for _, r := range tangle.GetTransactionHashesForAddress(hornet.Hash(addressHash), query.ValueOnly, true, maxResults-len(results)) {
+				results[string(r)] = struct{}{}
+			}
+		}
+		searchedBefore = true
+	} else {
+		// check if results match at least one of the address search criteria
+		for txHash := range results {
+			contains := false
+			for addressHash := range queryAddressHashes {
+				if tangle.ContainsAddress(hornet.Hash(addressHash), hornet.Hash(txHash), query.ValueOnly) {
+					contains = true
+					break
+				}
+			}
+			if !contains {
+				delete(results, txHash)
+			}
+		}
+	}
+
+	if !searchedBefore {
+		// search txs by tags
+		for tagHash := range queryTagHashes {
+			for _, r := range tangle.GetTagHashes(hornet.Hash(tagHash), true, maxResults-len(results)) {
+				results[string(r)] = struct{}{}
+			}
+		}
+	} else {
+		// check if results match at least one of the tag search criteria
+		for txHash := range results {
+			contains := false
+			for tagHash := range queryTagHashes {
+				if tangle.ContainsTag(hornet.Hash(tagHash), hornet.Hash(txHash)) {
+					contains = true
+					break
+				}
+			}
+			if !contains {
+				delete(results, txHash)
+			}
+		}
+	}
+
+	// convert to slice
+	var j int
+	txHashes := make([]string, len(results))
+	for r := range results {
+		txHashes[j] = hornet.Hash(r).Trytes()
+		j++
 	}
 
 	c.JSON(http.StatusOK, FindTransactionsReturn{Hashes: txHashes})