+++ CHANGELOG.md
@@ -2,6 +2,56 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.5.1] - 27.08.2020
+
+### Added
+
+    - getSpammerTips webapi call
+    - webapi route to control the spammer plugin
+    - autostart to the spammer config
+
+### Changed
+
+    - save memory in ApproversTraverser if walkAlreadyDiscovered is set
+    - findTransaction now returns an intersection of the search criteria
+    - show database size in GB if size > 1 GB (dashboard)
+    - moved dashboard frontend code to [another git repo](https://github.com/iotaledger/node-dashboard)
+    - spammer doesn't start automatically at node startup
+    - getNodeInfo now returns the connected peers count
+    - re-add logger settings to mainnet config file
+
+### Fixed
+
+    - warpsync milestone requesting
+    - OOM bug in the future cone solidifier
+    - map concurrent write/read panic in websocket
+    - bundle creation race condition in coordinator
+    - fix MQTT panic
+
+### Config file changes
+
+`config.json`
+
+```diff
+  "httpAPI": {
++    "permittedRoutes": [
++      "healthz"
++    ],
+  }
++  "logger": {
++    "level": "info",
++    "disableCaller": true,
++    "encoding": "console",
++    "outputPaths": [
++      "stdout"
++    ]
++  },
+  "spammer": {
+-    "semiLazyTipsLimit": 30
++    "autostart": false
+  },
+```
+
 ## [0.5.0] - 19.08.2020
 
 :warning: **Breaking change:** :warning:
@@ -41,10 +91,6 @@ The old HORNET versions won't be functional within the mainnet anymore!
     - Set higher default verticesLimit in visualizer
     - Update to Go 1.15
 
-### Removed
-
-    - Unused RefsInvalidBundles cache
-
 ### Fixed
 
     - Database revalidation
@@ -75,11 +121,11 @@ The old HORNET versions won't be functional within the mainnet anymore!
     - Monitor plugin
     - Legacy gossip protocol
     - Genesis tx special case
+    - RefsInvalidBundles cache
 
 ### Config file changes
 
 Please use the new config.json and transfer values from your current config.json over to the new one, as a lot of keys have changed or got removed (instead of mutating your current one).
-
 
 ## [0.4.2] - 22.07.2020